# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: getPurchases
      tags: [purchases]
      summary: List purchase invoices
      parameters:
        - name: supplier_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: date_from
          in: query
          schema:
            type: ["string", "null"]
        - name: date_to
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase invoice and update stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get purchase invoice details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseWithItems"
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: party_type
          in: query
          schema:
            type: string
            enum: [supplier, customer]
        - name: party_id
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /returns:
    get:
      operationId: getReturns
      tags: [returns]
      summary: List returns
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [sale, purchase]
      responses:
        "200":
          description: List of returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Return"
    post:
      operationId: createReturn
      tags: [returns]
      summary: Create a return
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Return"
...
  /reports/expiry-analysis:
    get:
      operationId: getExpiryAnalysis
      tags: [reports]
      summary: Get expiry analysis for medicines
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1month, 3months, 6months]
            default: 3months
      responses:
        "200":
          description: Expiry analysis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpiryAnalysisItem"
...
  /reports/top-selling:
    get:
      operationId: getTopSelling
      tags: [reports]
      summary: Top selling medicines
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top selling medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopSellingItem"
...
  /reports/daily-sales:
    get:
      operationId: getDailySalesReport
      tags: [reports]
      summary: Daily/monthly sales and profit summary
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, month, year]
            default: today
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSummary"
...
components:
  schemas:
...
    PurchaseItem:
      type: object
      required: [medicineId, medicineName, quantity, unitPrice, total]
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        total:
          type: number
        productionDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
    Purchase:
      type: object
      required: [id, invoiceNumber, total, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        total:
          type: number
        status:
          type: string
        createdAt:
          type: string
    PurchaseWithItems:
      type: object
      required: [id, invoiceNumber, total, createdAt, items]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        total:
          type: number
        status:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
    PurchaseItemInput:
      type: object
      required: [medicineId, quantity, unitPrice]
      properties:
        medicineId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        productionDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
    PurchaseInput:
      type: object
      required: [items]
      properties:
        supplierId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItemInput"
    Payment:
      type: object
      required: [id, partyType, amount, method, createdAt]
      properties:
        id:
          type: integer
        partyType:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: ["string", "null"]
        amount:
          type: number
        method:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [partyType, amount, method]
      properties:
        partyType:
          type: string
          enum: [supplier, customer]
        partyId:
          type: ["integer", "null"]
        amount:
          type: number
        method:
          type: string
          enum: [cash, check, transfer]
        notes:
          type: ["string", "null"]
    Return:
      type: object
      required: [id, type, medicineId, medicineName, quantity, amount, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        medicineId:
          type: integer
        medicineName:
          type: string
        referenceId:
          type: ["integer", "null"]
        quantity:
          type: integer
        amount:
          type: number
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
    ReturnInput:
      type: object
      required: [type, medicineId, quantity, amount]
      properties:
        type:
          type: string
          enum: [sale, purchase]
        medicineId:
          type: integer
        referenceId:
          type: ["integer", "null"]
        quantity:
          type: integer
        amount:
          type: number
        reason:
          type: ["string", "null"]
...
    ExpiryAnalysisItem:
      type: object
      required: [id, name, expiryDate, quantity, daysUntilExpiry]
      properties:
        id:
          type: integer
        name:
          type: string
        barcode:
          type: ["string", "null"]
        expiryDate:
          type: string
        quantity:
          type: integer
        daysUntilExpiry:
          type: integer
        urgency:
          type: string
...
    TopSellingItem:
      type: object
      required: [medicineId, medicineName, totalQuantity, totalRevenue]
      properties:
        medicineId:
          type: integer
        medicineName:
          type: string
        totalQuantity:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: number
...
    SalesSummary:
      type: object
      required: [totalSales, totalProfit, totalExpenses, invoiceCount]
      properties:
        totalSales:
          type: number
        totalProfit:
          type: number
        totalExpenses:
          type: number
        invoiceCount:
          type: integer
        period:
          type: string
...